requirements.txt
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
pillow==10.3.0
tqdm==4.66.2
timm==0.9.16
flask==3.0.3
jupyter==1.0.0
